refactor: rename project to traditional_saju
