+++ .fvmrc
@@ -1,3 +1,3 @@
 {
-  "flutter": "3.35.4"
+  "flutter": "stable"
 }
