+++ README.md
@@ -1,4 +1,4 @@
-# Traditional Saju V1
+# Traditional Saju
 
 ![coverage][coverage_badge]
 [![style: very good analysis][very_good_analysis_badge]][very_good_analysis_link]
@@ -31,7 +31,7 @@ $ flutter run --flavor staging --target lib/main_staging.dart
 $ flutter run --flavor production --target lib/main_production.dart
 ```
 
-_\*Traditional Saju V1 works on iOS, Android, Web, and Windows._
+_\*Traditional Saju works on iOS, Android, Web, and Windows._
 
 ---
 
@@ -92,7 +92,7 @@ This project relies on [flutter_localizations][flutter_localizations_link] and f
 3. Use the new string
 
 ```dart
-import 'package:traditional_saju_v1/l10n/l10n.dart';
+import 'package:traditional_saju/l10n/l10n.dart';
 
 @override
 Widget build(BuildContext context) {
