+++ analysis_options.yaml
@@ -1,7 +1,7 @@
 include: package:very_good_analysis/analysis_options.yaml
 analyzer:
   exclude:
-    - lib/l10n/gen/*
+    - lib/presentation/l10n
 linter:
   rules:
     public_member_api_docs: false
