+++ android/app/build.gradle.kts
@@ -14,7 +14,7 @@ if (keystorePropertiesFile.exists()) {
 }
 
 android {
-    namespace = "com.mandacode.traditional_saju_v1"
+    namespace = "com.mandacode.traditional_saju"
     compileSdk = flutter.compileSdkVersion
     ndkVersion = flutter.ndkVersion
 
@@ -29,7 +29,7 @@ android {
 
     defaultConfig {
         // TODO: Specify your own unique Application ID (https://developer.android.com/studio/build/application-id.html).
-        applicationId = "com.mandacode.traditional_saju_v1"
+        applicationId = "com.mandacode.traditional_saju"
         // You can update the following values to match your application needs.
         // For more information, see: https://flutter.dev/to/review-gradle-config.
         minSdk = flutter.minSdkVersion
@@ -60,17 +60,17 @@ android {
         create("production") {
             dimension = "default"
             applicationIdSuffix = ""
-            manifestPlaceholders["appName"] = "Traditional Saju V1"
+            manifestPlaceholders["appName"] = "Traditional Saju"
         }
         create("staging") {
             dimension = "default"
             applicationIdSuffix = ".stg"
-            manifestPlaceholders["appName"] = "[STG] Traditional Saju V1"
+            manifestPlaceholders["appName"] = "[STG] Traditional Saju"
         }
         create("development") {
             dimension = "default"
             applicationIdSuffix = ".dev"
-            manifestPlaceholders["appName"] = "[DEV] Traditional Saju V1"
+            manifestPlaceholders["appName"] = "[DEV] Traditional Saju"
         }
     }
