+++ macos/Runner/Configs/AppInfo.xcconfig
@@ -5,10 +5,10 @@
 // 'flutter create' template.
 
 // The application's name. By default this is also the title of the Flutter window.
-PRODUCT_NAME = my_app
+PRODUCT_NAME = traditional_saju
 
 // The application's bundle identifier
-PRODUCT_BUNDLE_IDENTIFIER = com.example.myApp
+PRODUCT_BUNDLE_IDENTIFIER = com.mandacode.traditional_saju
 
 // The copyright displayed in application information
 PRODUCT_COPYRIGHT = Copyright © 2023 com.example. All rights reserved.
