--- test/app/view/app_test.dart
@@ -1,15 +1,0 @@
-// Ignore for testing purposes
-// ignore_for_file: prefer_const_constructors
-
-import 'package:flutter_test/flutter_test.dart';
-import 'package:traditional_saju_v1/app/app.dart';
-import 'package:traditional_saju_v1/counter/counter.dart';
-
-void main() {
-  group('App', () {
-    testWidgets('renders CounterPage', (tester) async {
-      await tester.pumpWidget(App());
-      expect(find.byType(CounterPage), findsOneWidget);
-    });
-  });
-}
