--- test/counter/cubit/counter_cubit_test.dart
@@ -1,26 +1,0 @@
-import 'package:bloc_test/bloc_test.dart';
-import 'package:flutter_test/flutter_test.dart';
-
-import 'package:traditional_saju_v1/counter/counter.dart';
-
-void main() {
-  group('CounterCubit', () {
-    test('initial state is 0', () {
-      expect(CounterCubit().state, equals(0));
-    });
-
-    blocTest<CounterCubit, int>(
-      'emits [1] when increment is called',
-      build: CounterCubit.new,
-      act: (cubit) => cubit.increment(),
-      expect: () => [equals(1)],
-    );
-
-    blocTest<CounterCubit, int>(
-      'emits [-1] when decrement is called',
-      build: CounterCubit.new,
-      act: (cubit) => cubit.decrement(),
-      expect: () => [equals(-1)],
-    );
-  });
-}
