+++ web/index.html
@@ -29,7 +29,7 @@
   <!-- Favicon -->
   <link rel="icon" type="image/png" href="favicon.png"/>
 
-  <title>Traditional Saju V1</title>
+  <title>traditional_saju</title>
   <link rel="manifest" href="manifest.json">
 </head>
 <body>
