+++ windows/runner/Runner.rc
@@ -89,13 +89,13 @@ BEGIN
     BEGIN
         BLOCK "040904e4"
         BEGIN
-            VALUE "CompanyName", "com.mandacode.traditional_saju_v1" "\0"
-            VALUE "FileDescription", "traditional_saju_v1" "\0"
+            VALUE "CompanyName", "com.mandacode.traditional_saju" "\0"
+            VALUE "FileDescription", "traditional_saju" "\0"
             VALUE "FileVersion", VERSION_AS_STRING "\0"
-            VALUE "InternalName", "traditional_saju_v1" "\0"
-            VALUE "LegalCopyright", "Copyright (C) 2025 com.mandacode.traditional_saju_v1. All rights reserved." "\0"
-            VALUE "OriginalFilename", "traditional_saju_v1.exe" "\0"
-            VALUE "ProductName", "Traditional Saju V1" "\0"
+            VALUE "InternalName", "traditional_saju" "\0"
+            VALUE "LegalCopyright", "Copyright (C) 2025 com.mandacode.traditional_saju. All rights reserved." "\0"
+            VALUE "OriginalFilename", "traditional_saju.exe" "\0"
+            VALUE "ProductName", "Traditional Saju" "\0"
             VALUE "ProductVersion", VERSION_AS_STRING "\0"
         END
     END
